+++ src/redturtle/volto/upgrades.py
@@ -635,3 +635,7 @@ def to_4600(context):
     portal_types = api.portal.get_tool(name="portal_types")
     portal_types["Plone Site"].default_view = "homepage_view"
     portal_types["Plone Site"].immediate_view = "homepage_view"
+
+
+def to_4700(context):
+    installOrReinstallProduct(api.portal.get(), "collective.searchblocks")