+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Fix bug search event restapi (#148)
   [mamico]
-
+- Make AdvancedQuery ranking rules configurable.
+  [cekk]
 
 5.9.3 (2025-11-24)
 ------------------